Renaming and moving some files into something more managable. - adding db dir - adding lib dir
Refactor init file, and let lib/config.py handle the hasle.
This refactor include the ability to load different .ini config file.
The production's and test's .ini files would be handled smoothly.

Try to make the test suite as portable as possible, so that later
in automation testing we do it on the fly, without touching the
real database.