Adding some auth test: - Verify token test - Refresh token test I learnt how to use pickle at last, lol.
Refactor authentication to tokenizer. And make it singleton on init.py
to reduce object redundancy.

Move route of /verify and /refresh into /auth/verify and /auth/refresh.

Modify RBAC policy so that accessing /auth/verify and /auth/refresh
is available for anyone, aka @passthrough is used, but this is
worth noting since it bring more insecure checks..